requirements.txt
fastapi
uvicorn
openai
langchain
chromadb
pypdf
python-dotenv
tiktoken 
langchain-community
langchain-openai
langchain_chroma
python-multipart
# pytest
# pytest-asyncio